Reworked Underpowered Government Civics
Citizenship: Removed building production modifiers, added Rush Units with Gold*
Republic: Removed Rush Buildings with Gold*, and Food Penalties
Elective: Removed 2 Gold from Camps, Pastures, Palace production modifier, and 1 commerce on unimproved tiles, added 2 Food from Pastures, Upkeep is now None

Hannibal's favorite civic has been switched from Merchant Trade to Republic to ensure he doesn't switch to Monarchy.

Citizenship now reads: +2 Happiness in 4 largest cities, can rush units with gold*
Republic now reads: +1 Food per Specialist
Elective no reads: +2 Food from Pasture

Citizenship was changed to be less of a civic for getting your empire up and running and more of a civic for small or centralized nations with a great economy.

Republic was changed to be less punishing. Previously it's Specialist effect was far too often overshadowed by it's Food penalties, creating a trade off between Farms and Specialist. Even the nations that would have wanted to use the civic often wouldn't under player control due to the greater benefits of other civics and the lack of population and specialist slots throughout a majority of the early game. This new effect is a lot more simple, a single line effect like Despotism, Monarchy, and Democracy are. There's no longer a confusing list of improvements that are penalized.

Neither is there rushing units with gold*, which is typically not very useful for extended periods of time due to the high cost of units and low profits in the early game, which has led to Republic being a civic that players would switch to and then switch off of 3-5 turns later, once they get the units they wanted. Placing this effect on Citizenship ironically helps it be a civic for extended use. Due to the lack of competing civics during the early game and Vassalage's existence as a sidegrade, and not explicitly superior, Citizenship isn't optimal to switch to and off of in short succession like DOC Republic and as such players are able to wait to build up enough gold for another purchase.

Elective was probably the most troubling civic to buff. It's existing effects couldn't be buffed without becoming OP or too situational. Buffing palace build times would make it a on and off switch civic only used for a rapid build of a palace, something players would rarely do anyways. Buffing unimproved tile yields would make it a civic for when you don't have many if any improvements, and once you get enough workers to roughly match your population growth rates it becomes largely irrelevant. From my experience, the best strategy is to get as many tiles as you can improved as fast as you can by making as many workers as you can as soon as you can, without making too many. This strategy ensures you can get your nation running asap and thus you can begin your war machine sooner. Buffing unimproved tile yields would only serve to buff a playstyle that is not very powerful as it is. Buffing the commerce on pastures and camps meanwhile would quickly become OP the more commerce is added, with 3 or 4 already being too much IMO. Buffing camps also exclusively helps Holy Rome, with Poland, Mali, and Holland getting little benefit.

Giving 2 Food to Pastures ended up being the best addition. In my testing with the AI, the only nations that take it are Holy Rome and Poland, meaning that it's currently too UP in the eyes of the AI to pick. However I feel that this effect is strong enough to warrant use by players, which means it fits that sweet spot of not being powerful enough for everyone to use, but also not being weak enough that players won't want to use it historically. I've noticed it may also be good in the hands of the Scandinavians, English, and Frence especially, which fits well with it's limited historical use by those nations (according to wikipedia, though the Nordic portion is uncited I've read about the Thing before, though IIRC it was referred to as a Classical Republic then)

TL;DR Republic and Elective have been buffed, Citizenship has been given a sidegrade that makes it less of an obvious choice for new civs.

* I am using a shorthand version of the actual effect for the sake of brevity